Returned StubHandlingMiddleware.cs to old situation
