--- src/HttPlaceholder.Application/StubExecution/Commands/HandleStubRequest/HandleStubRequestCommand.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-
-namespace HttPlaceholder.Application.StubExecution.Commands.HandleStubRequest;
-
-/// <summary>
-///     A command for handling a stub request.
-/// </summary>
-public class HandleStubRequestCommand : IRequest<Unit>
-{
-}
